dom2: Implement CharacterData and friends
This implements the CharacterData methods as well as fill out the
simple classes that directly inherit from CharacterData.

I'm not sure what the deal with ProcessingInstruction is. The spec
specifies a member and a getter, but no setter or constructor, so we'll
just have to see how it's used in the rest of the spec when we get there
I suppose.