ci: Drop compile_command.json checks
I didn't have to do anything to get it working when adding it, it's
never failed since, and it almost doubles the time it takes our Windows
jobs to run.

The jobs sort of served like documentation for how to use it, but now
that it's added to the README, that's no longer needed.