meta: Set up a config file for gitlint
This works on all but 3 commits made during the past year. It's a bit
too lenient, in that any value after the '/' used in e.g. 'util/'
commits is valid, but we'll have to manually review that part for now.

The 3 commits failing:
* 1a1d177589: leading whitespace in title
* cacae1fc9e: starts with docs instead of doc
* 5a34fcbe3b: commit title is too long

Checked with `gitlint --commits fdac97f^..e66c14e`.