browser/gui: Handle CSS special backgrounds
If the html or body element has a non-transparent background set, that
should be used as the canvas background and not just as the element
background.

This is currently done in hacky way outside of //render as doing
correctly, not painting over things outside of the web page, without
hacks in //render will require //gfx functionality not in place yet.