uri: Migrate back to std::regex due to ctre stack-size issues
I tried adjusting the regex to make captures less greedy, but I wasn't
able to get the regex to both be correct and not cause stack overflows
when using clang-cl.

I'm leaving the ctre dependency set up as we'll probably want to use it
for simpler regex tasks like searching in a web page, where we won't
need any captures and only need to find where a match is.