ci: Re-enable the clang-cl job
We do this by switching to a Bazel release candidate as that's the only
option that supports clang-cl-16's include paths.

The timeout for the clang-cl job was also increased a bit as it timed
out due to waiting an extra 10 minutes before getting any output from
Bazel and the build actually showed any progress.