dom2: Remove stale beginning of a spec-compliant dom implementation
This hasn't really been touched since it was stubbed out in 2021 except
for clang-tidy warnings being fixed as they're enabled.

I don't think I'd implement this the same way if I were to start it
today, but even if we were to decide we love this old design, it's just
one revert away.