deps: Update hedron_compile_commands
This is needed for us to be able to generate compile commands on Windows
due to missing argument file support (I think) in the old version we
were previously using.

Generating compile commands for Linux is still pretty broken due to
missing support for the parse_headers-feature some of our dependencies
use.